# packages/api/openapi.yaml
openapi: 3.0.0
info:
  title: Field Nurse Schedule API
  version: 2.0.0
  description: API for managing field nurse scheduling

servers:
  - url: https://fns-scheduler-app-095cc2bbe204.herokuapp.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Local development server

components:
  schemas:
    TimeSlot:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the time slot
        start_time:
          type: string
          format: date-time
          description: Start time of the slot
        end_time:
          type: string
          format: date-time
          description: End time of the slot
        status:
          type: string
          enum: [available, reserved]
          description: Current status of the slot
        nurse_id:
          type: string
          nullable: true
          description: ID of the nurse who reserved the slot
        created_at:
          type: string
          format: date-time
          description: When the slot was created
        updated_at:
          type: string
          format: date-time
          description: When the slot was last updated
      required:
        - id
        - start_time
        - end_time
        - status

    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error

paths:
  /health:
    get:
      summary: Health check endpoint
      description: Returns the health status of the API
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [ok]
                required:
                  - status

  /slots:
    get:
      summary: Get all slots
      description: Retrieve all time slots with optional status filter
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [available, reserved]
          description: Filter slots by status
      responses:
        '200':
          description: List of time slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimeSlot'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /slots/{date}:
    get:
      summary: Get slots for a specific date
      description: Retrieve all time slots for a specific date
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
          description: Date in YYYY-MM-DD format
      responses:
        '200':
          description: List of time slots for the specified date
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimeSlot'
        '400':
          description: Invalid date format or date too far in the future
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /slots/{id}/reserve:
    post:
      summary: Reserve a time slot
      description: Reserve a time slot by either its ID or start_time
      parameters:
        - name: id
          in: path
          description: ID of the slot to reserve (can be 0 if using start_time)
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                start_time:
                  type: string
                  format: date-time
                  description: Start time of the slot to reserve (alternative to using ID)
              oneOf:
                - required: ['start_time']
                - {}
      responses:
        '200':
          description: Slot successfully reserved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeSlot'
        '400':
          description: Invalid request or slot not available
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum: 
                      - 'Either slot ID or start_time is required'
                      - 'Slot not available'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum: ['Too many requests']
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /slots/{id}/cancel:
    post:
      summary: Cancel a reservation
      description: Cancel a nurse's reservation for a time slot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID of the time slot to cancel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nurse_id:
                  type: string
                  description: ID of the nurse who made the reservation
              required:
                - nurse_id
      responses:
        '200':
          description: Reservation cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeSlot'
        '400':
          description: Invalid request or reservation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /slots/reset:
    post:
      summary: Reset all reserved slots
      description: Reset all reserved slots back to available status
      responses:
        '200':
          description: All slots reset successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Success message
                  updatedSlots:
                    type: array
                    items:
                      $ref: '#/components/schemas/TimeSlot'
                    description: List of slots that were reset
                required:
                  - message
                  - updatedSlots
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /slots/next-week:
    post:
      summary: Get available slots for upcoming days
      description: Retrieve all available time slots from now until a specified number of days in the future (maximum 10 days)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                days:
                  type: integer
                  minimum: 1
                  maximum: 10
                  description: Number of days to look ahead
              required:
                - days
      responses:
        '200':
          description: List of available time slots for the specified number of days
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimeSlot'
        '400':
          description: Invalid days parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'